Uploading profile picture by image service
