+++ src/main/java/com/example/vrs/controller/UserController.java
@@ -33,14 +33,5 @@ public class UserController {
 		
 	}
 	
-	@PostMapping("/upload-profile")
-	public ResponseEntity<SimpleResponseStructure> uploadProfile(@RequestParam ("userId") int userId,
-				@RequestParam("file") MultipartFile file ){
-		
-		userService.uploadProfile(userId,file);
-
-		return ResponseEntity.status(HttpStatus.CREATED)
-				.body(SimpleResponseStructure.create(HttpStatus.CREATED.value(), "Profile Picture Updated"));
-	}
 
 }
